requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
onnxruntime==1.16.3
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
